+++ utipy/array/__init__.py
@@ -2,6 +2,7 @@
 
 from .blend import blend
 from .windowed_reverse import windowed_reverse
+from .window import window